==> 4. Draw the links.rmd <==
---
title: "Draw the links"
author: "Rossella Alfano, Brigitte Reimann, Congrong Wang and Michelle Plusquin"
output:
  word_document: default
  html_document: default
---

```{r setup, include=FALSE}

require("knitr")
opts_knit$set(root.dir ="G:/My Drive/microbiome_workshop/data/")
knitr::opts_chunk$set(echo = TRUE)
```

In this tutorial, we will write a script to generate the links of a circular plot to display the correlations contained in a multi-omics matrix. 

You can download the data you need for this tutorial [here](link). This folder contains three sub-folders: 

- the “multi_omics-matrix” folder including the multi-omics matrix, 

- the “annotation” folder where are stored the annotation file needed in the tutoral,

- the “circos_files” folder in which you can find already the outputs of this tutorial. 

Please, change the data path in Rstudio “G:/My Drive/microbiome_workshop/data/" to the location where you stored the "multi_omics", "annotations" and "circos_files" folders.

# Code to draw the ideogram

In the script below, change the data path "G:/My Drive/microbiome_workshop/data/" to the location where the folders "multi_omics", "annotation" and "circos_files" and load the multi-omics matrix and the annotation files. 


```{r warning = FALSE,message = FALSE}
#####################Draw the links#####################
rm(list = ls())
##set your working directory here
setwd("G:/My Drive/microbiome_workshop/data/")#change the path to the one where you stored the folders "multi_omics", "annotation" and "circos_files" 
##read the multiomics matrix
readRDS("multi_omics_matrix/multiomics_matrix.rds")->dataset

##read the annotation
readRDS("annotation/ideogram_annotation.rds")->annot

```

Run the function below on the annotation file and multi-omics matrix to generate the links of your circular plot. 

```{r warning = FALSE,message = FALSE}
#write the fuction to draw the link
writeLinks <- function(dataset, annot, fileout="circos_files/links.txt") {
   cat("",file=fileout)
  for(i in 1:nrow(dataset)) {
    coord1 <-  subset(annot, annot$chr == dataset[i, 'omic_layer1'] & annot$Name == dataset[i, 'omic1'])$coord[1]
    coord2 <- subset(annot, annot$chr == dataset[i, 'omic_layer2'] & annot$Name == dataset[i, 'omic2'])$coord[1]
    txt1 <- paste(dataset[i, 'omic_layer1'], coord1-1, coord1)
    txt2 <- paste(dataset[i, 'omic_layer2'], coord2-1, coord2)
    optionString <- sprintf(" corr=%f", dataset[i, 'corr'])
    cat(sprintf("%s %s%s\n", txt1, txt2, optionString),file=fileout,append=T)
  }
}

##check they are characters
writeLinks(dataset,annot)

```

Now the links.txt is ready for use. 

You can [Run Circos](link) to plot the ideogram with links!
